Move the datadog package to appmetrics/emitter
Leave the old package in place but use type aliases and forwarding
functions to reference the new package. Mark all the existing types as
deprecated.
